moved spawnPhantom to separate file
